Safari-611.1.10/LayoutTests/fast/mediastream/MediaStream-page-muted-expected.txt

Test enabling/disabling mock media capture devices

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



*** Mock capture devices should be enabled by default
PASS mediaStream is an instance of Object
PASS mediaStream.getTracks().length is 2
PASS window.internals.pageMediaState().includes('HasActiveAudioCaptureDevice') && window.internals.pageMediaState().includes('HasActiveVideoCaptureDevice') became true

*** Muting capture devices
EVENT: mute
PASS muteChangedEvent.target.muted is true
EVENT: mute
PASS muteChangedEvent.target.muted is true

PASS window.internals.pageMediaState().includes(HasMutedAudioCaptureDevice) became true
PASS window.internals.pageMediaState().includes(HasMutedVideoCaptureDevice) became true
PASS window.internals.pageMediaState().includes(HasActiveAudioCaptureDevice) became false
PASS window.internals.pageMediaState().includes(HasActiveVideoCaptureDevice) became false

*** Unmuting capture devices
EVENT: unmute
PASS muteChangedEvent.target.muted is false
EVENT: unmute
PASS muteChangedEvent.target.muted is false

PASS window.internals.pageMediaState().includes(HasActiveAudioCaptureDevice) became true
PASS window.internals.pageMediaState().includes(HasActiveVideoCaptureDevice) became true
PASS window.internals.pageMediaState().includes(HasMutedAudioCaptureDevice) became false
PASS window.internals.pageMediaState().includes(HasMutedVideoCaptureDevice) became false
PASS successfullyParsed is true

TEST COMPLETE
